BackEnd/requirements.txt
absl-py==2.2.2
astunparse==1.6.3
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.0
flatbuffers==25.2.10
gast==0.6.0
google-auth==2.40.1
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
keras==2.15.0
libclang==18.1.1
Markdown==3.8
MarkupSafe==3.0.2
ml-dtypes==0.2.0
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==25.0
pillow==11.2.1
protobuf==4.25.7
psycopg==3.2.9
psycopg-binary==3.2.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
soupsieve==2.7
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.14.1
scipy==1.15.3


# py -3.11 -m venv .venv
# pip install -r requirements.txt